Remove unneeded type conversion

+++ bpf.go
@@ -217,7 +217,7 @@ func filterAttribute(filters []ConnAttr) []bpf.RawInstruction {
 	if nested != 0 {
 		for _, nest := range filterCheck[filters[0].Type].nest {
 			// find nest attribute
-			tmp = bpf.RawInstruction{Op: bpfLDX | bpfIMM, K: uint32(nest)}
+			tmp = bpf.RawInstruction{Op: bpfLDX | bpfIMM, K: nest}
 			raw = append(raw, tmp)
 			tmp = bpf.RawInstruction{Op: bpfLD | bpfB | bpfABS, K: 0xfffff00c}
 			raw = append(raw, tmp)
@@ -282,7 +282,7 @@ func filterSubsys(subsys uint32) []bpf.RawInstruction {
 	raw = append(raw, tmp)
 
 	// A == subsys ? jump + 1 : accept
-	tmp = bpf.RawInstruction{Op: bpfJMP | bpfJEQ | bpfK, Jt: 1, K: uint32(subsys)}
+	tmp = bpf.RawInstruction{Op: bpfJMP | bpfJEQ | bpfK, Jt: 1, K: subsys}
 	raw = append(raw, tmp)
 
 	// verdict -> accept
